Add common subproject
Move a bunch of generic libraries from net.rptools.lib into this project. Not everything could be moved over
immediately, though.
