+++ build.gradle
@@ -361,6 +361,7 @@ def zipJpackageImage = tasks.register('zipJpackageImage', Zip) {
 
 // In this section you declare the dependencies for your production and test code
 dependencies {
+    implementation project(':common')
     implementation project(':clientserver')
 
     implementation(libs.bundles.log4j)
