+++ common/src/main/java/net/rptools/lib/CollectionUtil.java
