+++ common/src/main/java/net/rptools/lib/DebugStream.java
