+++ common/src/main/java/net/rptools/lib/MathUtil.java
