+++ common/src/main/java/net/rptools/lib/OsDetection.java
