+++ common/src/main/java/net/rptools/lib/StringUtil.java
