+++ common/src/main/java/net/rptools/lib/image/RenderQuality.java
