+++ common/src/main/java/net/rptools/lib/image/ThumbnailManager.java
