+++ common/src/main/java/net/rptools/lib/net/SyrinscapeConnection.java
