+++ common/src/main/java/net/rptools/lib/net/SyrinscapeURLStreamHandler.java
