+++ settings.gradle
@@ -16,4 +16,4 @@ include 'services:webservice'
 */
 
 rootProject.name = 'MapTool'
-include ':clientserver'
+include 'common', ':clientserver'
